Fix node-runner failing on container restart

+++ docker/node-runner/startup.bash
@@ -2,6 +2,12 @@
 
 set -xe
 
+if [ -d "/app" ]; then
+  cd /app
+  npm run start
+  exit
+fi
+
 git clone $GIT_CLONE_URL /app
 cd /app
 git reset --hard $GIT_VERSION_HASH
